+++ service/not_found.go
@@ -7,13 +7,25 @@ import (
 	"github.com/hoop33/limo/model"
 )
 
+var errNotFound = errors.New("service not found")
+
 // NotFound is used when the specified service is not found
 type NotFound struct {
 }
 
 // Login is not implemented
 func (nf *NotFound) Login(ctx context.Context) (string, error) {
-	return "", errors.New("service not found")
+	return "", errNotFound
+}
+
+// AddStar is not implemented
+func (nf *NotFound) AddStar(ctx context.Context, token, owner, repo string) (*model.Star, error) {
+	return nil, errNotFound
+}
+
+// DeleteStar is not implemented
+func (nf *NotFound) DeleteStar(ctx context.Context, token, owner, repo string) (*model.Star, error) {
+	return nil, errNotFound
 }
 
 // GetStars is not implemented
@@ -26,4 +38,8 @@ func (nf *NotFound) GetTrending(ctx context.Context, trendingChan chan<- *model.
 
 // GetEvents is not implemented
 func (nf *NotFound) GetEvents(ctx context.Context, eventChan chan<- *model.EventResult, token, user string, page, count int) {
+}
+
+// SetInsecure sets the service to skip cert verification
+func (nf *NotFound) SetInsecure(insecure bool) {
 }